+++ src/wintext.c
@@ -1,5 +1,5 @@
 // wintext.c (part of mintty)
-// Copyright 2008-22 Andy Koppe, 2015-2024 Thomas Wolff
+// Copyright 2008-22 Andy Koppe, 2015-2025 Thomas Wolff
 // Adapted from code from PuTTY-0.60 by Simon Tatham and team.
 // Licensed under the terms of the GNU General Public License v3 or later.
 
@@ -692,8 +692,10 @@ win_init_fontfamily(HDC dc, int findex)
   GetObject(ff->fonts[FONT_NORMAL], sizeof(LOGFONT), &logfont);
   trace_font(("created font %s %d it %d cs %d\n", logfont.lfFaceName, (int)logfont.lfWeight, logfont.lfItalic, logfont.lfCharSet));
   SelectObject(dc, ff->fonts[FONT_NORMAL]);
-  GetTextMetrics(dc, &tm);
-  if (!tm.tmHeight) {
+
+  int tmok = GetTextMetrics(dc, &tm);
+  //printf("GTM %d h %d a %d d %d e %d i %d w %d cs %d\n", tmok, tm.tmHeight, tm.tmAscent, tm.tmDescent, tm.tmExternalLeading, tm.tmInternalLeading, tm.tmAveCharWidth, tm.tmCharSet);
+  if (!tmok || !tm.tmHeight) {
     // corrupt font installation (e.g. deleted font file)
     font_warning(ff, _("Font installation corrupt, using system substitute"));
     wstrset(&ff->name, W(""));
@@ -702,6 +704,11 @@ win_init_fontfamily(HDC dc, int findex)
     SelectObject(dc, ff->fonts[FONT_NORMAL]);
     GetTextMetrics(dc, &tm);
   }
+  // fix broken font metrics
+  if (tm.tmAveCharWidth < 0)
+    // Panic Sans reports negative char width
+    tm.tmAveCharWidth = - tm.tmAveCharWidth;
+
   if (!findex)
     lfont = logfont;
 
@@ -3360,8 +3367,11 @@ win_text(int tx, int ty,wchar *text, int len, cattr attr, cattr *textattr, ushor
     origtext = text;
     text = newn(wchar, len);
     // clear font glyphs under self-drawn geometric symbols
+    // - this method of clearing background is no longer in use since 3.7.8
+    // - keeping it in for now just in case; should be cleaned up later
     for (int i = 0; i < len; i++)
       text[i] = ' ';
+      //text[i] = ' ';
   }
 
  /* Array with offsets between neighbouring characters */
@@ -3504,7 +3514,10 @@ win_text(int tx, int ty,wchar *text, int len, cattr attr, cattr *textattr, ushor
   bool underlaid = false;
   void clear_run() {
     if (!underlaid) {
-      ExtTextOutW(dc, xt, yt, eto_options | ETO_OPAQUE, &box0, W(" "), 1, dxs);
+      // clear background of current output chunk
+      HBRUSH bgb = CreateSolidBrush(bg);
+      FillRect(dc, &box, bgb);
+      DeleteObject(bgb);
 
       underlaid = true;
     }
@@ -3536,6 +3549,8 @@ win_text(int tx, int ty,wchar *text, int len, cattr attr, cattr *textattr, ushor
       underlaid = false;
 #endif
   }
+  else if (origtext && !ldisp2)
+    clear_run();  // clear background for self-drawn characters (#1310)
 
  /* Coordinate transformation per line */
   int coord_transformed = 0;
@@ -3720,6 +3735,7 @@ draw:;
  /* Underline */
   if (!ldisp2 && lattr != LATTR_TOP &&
       (force_manual_underline ||
+       nfont & FONT_UNDERLINE ||  // ensure underline of self-drawn characters
        (attr.attr & (ATTR_DOUBLYUND | ATTR_BROKENUND)) ||
        ((attr.attr & UNDER_MASK) == ATTR_UNDER &&
         (ff->und_mode == UND_LINE || (attr.attr & ATTR_ULCOLOUR)))
@@ -3772,7 +3788,7 @@ draw:;
   if (ldisp1) {
     if (!underlaid)
       clear_run();
-    goto _return;
+    goto _return;  // skipping coord_transformed2 set and restore
   }
 
  /* Partial glyph adjustments */
@@ -3837,10 +3853,15 @@ draw:;
   if (combining || combining_double)
     *dxs = char_width;  // convince Windows to apply font underlining
 
- /* Now, really draw the text */
   // handle invisible and blinking attributes on image background
   if (fg == bg && default_bg && cfg.backgfile.type)
+    goto skip_drawing;  // restore coord_transformed2, then skip self-drawing
+
+  // skip text output for self-drawn characters
+  if (origtext)
     goto skip_drawing;
+
+ /* Now, really draw the text */
 
   text_out_start(dc, text, len, dxs);
 
@@ -3928,7 +3949,14 @@ skip_drawing:;
     box2 = box2_;
   }
 
- /* Manual drawing of certain graphics */
+
+ /* Skip self-drawing to handle invisible attribute on image background */
+  if (fg == bg && default_bg && *cfg.background)
+    goto _return;
+
+
+ /* Self-drawn characters: manual drawing of certain graphics */
+
   // line_width already set above for DEC Tech adjustments
 #define dont_debug_vt100_line_drawing_chars
 #ifdef debug_vt100_line_drawing_chars
@@ -4501,8 +4529,10 @@ skip_drawing:;
 
  /* Strikeout */
   if ((attr.attr & ATTR_STRIKEOUT)
+      //&& !ldisp1
       && (cfg.underl_manual || cfg.underl_colour != (colour)-1
           || (attr.attr & ATTR_ULCOLOUR)
+          || origtext  // apply strikeout to self-drawn characters
          )
      )
   {